+++ packages/mailbox-service/src/benethos_mailbox_service/web/api/deps.py
@@ -16,7 +16,7 @@ from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
 from ...data.models import MessageFilter
 from ...data.models.messages import SEARCH_TEXT_PATTERN
 from ...domain.access import Access
-from ..services import Accounts, Auth, Discoverer, Mailbox, Users
+from ..services import Accounts, Auth, Discoverer, Mailbox, Users, Webhooks
 from ..urls import client_address
 
 __all__ = [
@@ -28,6 +28,7 @@ __all__ = [
     "Search",
     "Since",
     "Users",
+    "Webhooks",
     "authenticate",
 ]